feat: add pydantic model utils
