+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "agentic-demo"
 version = "0.1.0"
-description = "LangGraph Lecture Builder"
+description = "Pydantic AI Lecture Builder"
 authors = [
     {name = "Jolyon Suthers",email = "201621+fenrick@users.noreply.github.com"}
 ]
@@ -14,13 +14,6 @@ dependencies = [
     "fastapi",
     "httpx",
     "jsonschema",
-    "langchain-community",
-    "langchain-openai",
-    "langchain-perplexity",
-    "langchain",
-    "langgraph-sdk",
-    "langgraph",
-    "langsmith",
     "logfire",
     "loguru",
     "markdown",
@@ -38,7 +31,6 @@ dependencies = [
     "sse-starlette",
     "uvicorn[standard]",
     "weasyprint",
-    "langgraph-checkpoint-sqlite (>=2.0.11,<3.0.0)",
 ]
 
 [build-system]
