+++ src/agents/pedagogy_critic.py
@@ -98,7 +98,11 @@ def classify_bloom_level(text: str) -> str:
         else:
             provider = OpenAIProvider()
         model = OpenAIModel(settings.model_name, provider=provider)
-        agent = Agent(model=model, result_type=BloomResult, instructions=instructions)
+        agent = Agent(
+            model=model,
+            result_type=BloomResult,
+            instructions=instructions,
+        )  # type: ignore[call-overload]
         result = agent.run_sync(text)
         level = result.data.level.strip().lower()
         if level in BLOOM_LEVELS:
